Reporter/requirements.txt
Flask==1.1.1
flask-restplus==0.13.0
pandas==0.25.1
mysql-connector==2.2.9
pytest==5.1.3
pytest-flask==0.15.0
flask-cors==3.0.8   